- PSPAudioCodec support. - Runtime mesh loading.
git-svn-id: https://fuji.googlecode.com/svn/trunk@544 7e75b669-355e-e21c-2746-f45317fb21cf

+++ Fuji/Private/Include/MFPtrList.inl
@@ -22,7 +22,7 @@ inline void MFPtrList<T>::Init(const char *pGroupName, uint32 maxElements)
 		*ppMark = 0;
 
 		for(uint32 i = 0; i<maxElements; ++i)
-			*(++ppMark) = (T*)(0xdeadbeef);
+			*(++ppMark) = (T*)((size_t)0xdeadbeef);
 
 		*(++ppMark) = 0;
 	}
